Removed sleep function from touch controller
Removed sleep mode for the touch controller to allow for touch to wake functionality. The downside to this is instead of consuming 5uA when asleep it will consume 10uA in standby and if the screen is touched that isnt a finger it will wake. I think we need what the apple watch had before the always on display.